Update docker-entrypoint.sh
On my machine the php instance was not coming up[ because it was searching for the database. Finally I logged in to the terminal of the image and there I was able to execute the command 'php bin/console dbal:run-sql "SELECT 1"'. Hence, I believe php is missing from the test in the 'until' on line 40.